Create TranslationTimelineProvider
